examples/crewai-procurement-agent/pyproject.toml
[project]
name = "attestara-crewai-procurement-agent"
version = "0.1.0"
description = "Reference CrewAI agent demonstrating the full Attestara protocol lifecycle"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.80.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
